% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCP.R
\name{contEndpoint}
\alias{contEndpoint}
\title{get Endpoint for continuous data}
\usage{
contEndpoint(
  paov,
  pks,
  pnormal,
  phomogeneity,
  monotonicity,
  william,
  dunnett,
  dunn,
  jonckheere,
  procedure = "stepDown",
  doses = c("A", "B", "C", "D")
)
}
\arguments{
\item{paov}{}

\item{pks}{}

\item{pnormal}{}

\item{phomogeneity}{}

\item{monotonicity}{}

\item{william}{}

\item{dunnett}{}

\item{dunn}{}

\item{jonckheere}{}

\item{procedure}{}

\item{doses}{}
}
\value{

}
\description{
get Endpoint for continuous data
}
