% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECxZG.R
\name{ED.ZG}
\alias{ED.ZG}
\title{Calculating ED following the regulatrory ED definition.}
\usage{
\method{ED}{ZG}(
  object,
  respLev,
  maxEff = TRUE,
  trend = "Increase",
  range = "Percentage",
  CI = c("delta", "inv"),
  ...
)
}
\arguments{
\item{object}{fitted model using drm}

\item{respLev}{response level, the x of ECx.}

\item{maxEff}{Not used for now. maximum effect.}

\item{trend}{Increase of Decrease, whether the dose response is decreasing or increasing compared control.}

\item{range}{using Percentage, not used for now.}

\item{...}{other parameters for ED function}

\item{interval}{methods to calculate the confidence intervals!}
}
\value{
Back calculated regulatory ECx
}
\description{
Calculating ED following the regulatrory ED definition.
}
