% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCP.R
\name{monotonicityTest}
\alias{monotonicityTest}
\title{Testing Monocity}
\usage{
monotonicityTest(Data, Treatment, Response)
}
\arguments{
\item{Data}{}

\item{Treatment}{}

\item{Response}{}
}
\value{

}
\description{
Testing Monocity
}
