% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECxZG.R
\name{mselect.ZG}
\alias{mselect.ZG}
\title{Title}
\usage{
mselect.ZG(
  object = NULL,
  fctList = NULL,
  nested = FALSE,
  sorted = c("IC", "Res var", "Lack of fit", "no"),
  linreg = FALSE,
  icfct = AIC,
  respCol = "effect",
  doseCol = "dose",
  data = NULL,
  type = "continuous",
  EFSA = NULL
)
}
\arguments{
\item{EFSA}{}
}
\value{
mselect.ZG
}
\description{
Title
}
