% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECxZG.R
\name{plot.edList}
\alias{plot.edList}
\title{Plot the ECx estimation and confidence intervals from the list of models.}
\usage{
\method{plot}{edList}(edList, fctNames, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot the ECx estimation and confidence intervals from the list of models.
}
