% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECxZG.R
\name{plot.modList}
\alias{plot.modList}
\title{Plot a list of models together.}
\usage{
\method{plot}{modList}(
  modList,
  respLev = NULL,
  data = NULL,
  xmin,
  xmax,
  scale = c("logx", "logy", "logxy", "orig"),
  npts = 100,
  plot_respLev = FALSE,
  xbreaks = NULL,
  ymin = NULL,
  ymax = NULL,
  ...
)
}
\arguments{
\item{...}{user input when the use does not want to use the data provided in the modList}
}
\value{

}
\description{
Plot a list of models together.
}
