% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandling.R
\name{Raw2Long}
\alias{Raw2Long}
\title{Reading raw ToxRat/.. data according to Data type.}
\usage{
Raw2Long(path, sheet = "RawData Fresh Weight", DataType = "Myriophyllum")
}
\arguments{
\item{path}{path to the file (excel sheets)}

\item{sheet}{sheet of relevant endpoint}

\item{DataType}{either "Algae", "Myriophyllum", "Lemna","Earthworm", or ""}
}
\value{
a data table of long format, suitable for drm function
}
\description{
Reading raw ToxRat/.. data according to Data type.
}
